% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{CrossValidationResult-class}
\alias{CrossValidationResult}
\alias{CrossValidationResult-class}
\alias{class:CrossValidationResult}
\title{Cross Validation Result Class}
\description{
Cross Validation Result Class
}
\details{
Instances of this class store the result of cross
validation.
}
\section{Slots}{

\describe{
\item{\code{cross}}{number of folds for cross validation}

\item{\code{noCross}}{number of CV runs}

\item{\code{groupBy}}{group assignment of samples}

\item{\code{perfParameters}}{collected performance parameters}

\item{\code{outerCV}}{flag indicating outer CV}

\item{\code{folds}}{folds used in CV}

\item{\code{cvError}}{cross validation error}

\item{\code{foldErrors}}{fold errors}

\item{\code{noSV}}{number of support vectors}

\item{\code{ACC}}{cross validation accuracy}

\item{\code{BACC}}{cross validation balanced accuracy}

\item{\code{MCC}}{cross validation Matthews correlation coefficient}

\item{\code{AUC}}{cross validation area under the ROC curve}

\item{\code{foldACC}}{fold accuracy}

\item{\code{foldBACC}}{fold balanced accuracy}

\item{\code{foldMCC}}{fold Matthews correlation coefficient}

\item{\code{foldAUC}}{fold area under the ROC curve}

\item{\code{sumAlphas}}{sum of alphas}
}}

