% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{GappyPairKernel-class}
\alias{GappyPairKernel}
\alias{GappyPairKernel-class}
\alias{class:GappyPairKernel}
\title{Gappy Pair Kernel Class}
\description{
Gappy Pair Kernel Class
}
\details{
Instances of this class represent a kernel object for the
gappy pair kernel. The kernel considers adjacent pairs of
kmers with up to m irrelevant characters between the pair.
The class is derived from \link{SequenceKernel}.
}
\section{Slots}{

\describe{
\item{\code{k}}{length of the substrings considered by the kernel}

\item{\code{m}}{maximum number of irrelevant character between two kmers}

\item{\code{r}}{exponent (for details see \link{gappyPairKernel})}

\item{\code{annSpec}}{when set the kernel evaluates annotation information}

\item{\code{distWeight}}{distance weighting function or vector}

\item{\code{normalized}}{data generated with this kernel object is normalized}

\item{\code{exact}}{use exact character set for evaluation}

\item{\code{ignoreLower}}{ignore lower case characters in the sequence}

\item{\code{presence}}{consider only the presence of kmers not their counts}

\item{\code{revComplement}}{consider a kmer and its reverse complement
as the same feature}

\item{\code{mixCoef}}{mixing coefficients for mixture kernel}
}}

