% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{kebabsCollectInfo}
\alias{kebabsCollectInfo}
\title{Collect KeBABS Package Information}
\usage{
kebabsCollectInfo(onlyKebabsRelated = TRUE)
}
\arguments{
\item{onlyKebabsRelated}{if set to \code{TRUE} only the packages related
to KeBABS are shown, if set to \code{FALSE} all attached packages and
all packages loaded via namespace are shown. Default=TRUE}
}
\value{
see above
}
\description{
Collects and prints general R and package version information.
If you have a question related to some KeBABS functionality or observe
some unexpected behavior please call this function and send its output
together with your information to the contact address specified on the
title page of the package vignette. The function by default only outputs
the package version of those packages which are directly related to the
KeBABS functionality.
}
\examples{
## collect KeBABS related package information
kebabsCollectInfo()
}
\author{
Johannes Palme <kebabs@bioinf.jku.at>
}

