% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{KernelMatrix-class}
\alias{KernelMatrix}
\alias{KernelMatrix-class}
\alias{class:KernelMatrix}
\title{Kernel Matrix Class}
\description{
Kernel Matrix Class
}
\details{
Instances of this class are used in KeBABS for storing a
kernel matrix. The hidden data part ".Data" contains the
matrix.
}

