% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MismatchKernel-class}
\alias{MismatchKernel}
\alias{MismatchKernel-class}
\alias{class:MismatchKernel}
\title{Mismatch Kernel Class}
\description{
Mismatch Kernel Class
}
\details{
Instances of this class represent a kernel object for the mismatch
kernel. The class is derived from \link{SequenceKernel}.
}
\section{Slots}{

\describe{
\item{\code{k}}{length of the substrings considered by the kernel}

\item{\code{m}}{maximum number of mismatches}

\item{\code{r}}{exponent (for details see \link{mismatchKernel})}

\item{\code{annSpec}}{not used for mismatch kernel}

\item{\code{distWeight}}{not used for mismatch kernel}

\item{\code{normalized}}{data generated with this kernel object is normalized}

\item{\code{exact}}{use exact character set for evaluation}

\item{\code{ignoreLower}}{ignore lower case characters in the sequence}

\item{\code{presence}}{consider only the presence of kmers not their counts}

\item{\code{revComplement}}{not used for mismatch kernel}

\item{\code{mixCoef}}{not used for mismatch kernel}
}}
\author{
Johannes Palme
}
\references{
\url{https://github.com/UBod/kebabs}\cr\cr
J. Palme, S. Hochreiter, and U. Bodenhofer (2015) KeBABS: an R package
for kernel-based analysis of biological sequences.
\emph{Bioinformatics}, 31(15):2574-2576.
DOI: \doi{10.1093/bioinformatics/btv176}.
}
