% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{ModelSelectionResultAccessors}
\alias{ModelSelectionResultAccessors}
\alias{gridColumns}
\alias{gridColumns,ModelSelectionResult-method}
\alias{gridErrors}
\alias{gridErrors,ModelSelectionResult-method}
\alias{gridRows}
\alias{gridRows,ModelSelectionResult-method}
\alias{performance}
\alias{performance,ModelSelectionResult-method}
\title{ModelSelectionResult Accessors}
\usage{
\S4method{gridRows}{ModelSelectionResult}(object)
}
\arguments{
  \item{object}{a model selection result object (can be
  extracted from KeBABS model with accessor
  \code{\link{modelSelResult}})}
}
\value{
\code{gridRows}: returns a list of kernel objects\cr
\code{gridColumns}: returns a \code{DataFrame} object with
grid column parameters\cr \code{gridErrors}: returns a
matrix with grid errors\cr \code{performance}: returns a
list of matrices with performance values
}
\description{
ModelSelectionResult Accessors
}
\section{Accessor-like methods}{
  \describe{ \item{}{\code{gridRows}: return the grid rows
  containing the kernels.  } \item{}{\code{gridColumns}:
  return the grid columns.  } \item{}{\code{gridErrors}:
  return the grid CV errors.  } \item{}{\code{performance}:
  return the collected performance parameters.  } }
}
\examples{
## create kernel object for normalized spectrum kernel
specK5 <- spectrumKernel(k=5)
\dontrun{
## load data
data(TFBS)

## perform training - feature weights are computed by default
model <- kbsvm(enhancerFB, yFB, specK5, pkg="LiblineaR",
               svm="C-svc", cost=c(1,15,50,100), cross=10,
               perfParameters="ALL", showProgress=TRUE)

## show model selection result
mres <- modelSelResult(model)
mres

## extract grid errors
gridErrors(mres)

## extract other performance parameters
performance(mres)
}
}

