% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{PredictionProfile-class}
\alias{PredictionProfile}
\alias{PredictionProfile-class}
\alias{class:PredictionProfile}
\title{Prediction Profile Class}
\description{
Prediction Profile Class
}
\details{
This class stores prediction profiles generated for a set
of biological sequences from a trained model. Prediction
profiles show the relevance of individual sequence
positions for the prediction result.
}
\section{Slots}{

\describe{
\item{\code{sequences}}{sequence information for the samples with profiles}

\item{\code{baselines}}{baselines generated from the offset in the model spread\cr
to all sequence positions}

\item{\code{profiles}}{prediction profile information stored as dense matrix with\cr
the rows as samples and the columns as positions in the sample}

\item{\code{kernel}}{kernel used for training the model on which these prediction\cr
profiles are based}
}}

