% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ROCData-class}
\alias{ROCData}
\alias{ROCData-class}
\alias{class:ROCData}
\title{ROC Data Class}
\description{
ROC Data Class
}
\details{
This class stores receiver operating characteristics (ROC)
data.
}
\section{Slots}{

\describe{
\item{\code{AUC}}{area under ROC curve}

\item{\code{TPR}}{true positive rate for varying threshold}

\item{\code{FPR}}{false positive rate for varying threshold}
}}

