% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runit}
\alias{runit}
\title{RUnit Tests for Kernels}
\usage{
runit(unit = NULL, echo = FALSE, silent = TRUE, seed = 789, ...)
}
\arguments{
  \item{unit}{just for internal testing purpose.}

  \item{echo}{just for internal testing purpose.}

  \item{silent}{just for internal testing purpose.}

  \item{seed}{just for internal testing purpose.}

  \item{...}{just for internal testing purpose.}
}
\value{
see above
}
\description{
Start Command for RUnit Tests
}
\details{
This command is just for internal testing purpose and is
exported only for simplifying developer testing.
}
\examples{
runit("MotifKernel")
}
\author{
Johannes Palme <kebabs@bioinf.jku.at>
}
\references{
\url{http://www.bioinf.jku.at/software/kebabs}
}

