% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SequenceKernel-class}
\alias{SequenceKernel}
\alias{SequenceKernel-class}
\alias{class:SequenceKernel}
\title{Sequence Kernel Class}
\description{
Sequence Kernel Class
}
\details{
This class represents the parent class for all sequence kernels. It is an
abstract class and must not be instantiated.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{the kernel function is stored in this slot. It is executed
when the kernel matrix is created through invoking the kernel object.}

\item{\code{.userDefKernel}}{indicates whether kernel is user defined or not.}
}}
\author{
Johannes Palme <kebabs@bioinf.jku.at>
}
\references{
\url{http://www.bioinf.jku.at/software/kebabs/}\cr\cr
J. Palme, S. Hochreiter, and U. Bodenhofer (2015) KeBABS: an R package
for kernel-based analysis of biological sequences.
\emph{Bioinformatics}, 31(15):2574-2576, 2015.
DOI: \doi{10.1093/bioinformatics/btv176}.
}

