% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{SVMInformation-class}
\alias{SVMInformation}
\alias{SVMInformation-class}
\alias{class:SVMInformation}
\title{SVM Information Class}
\description{
SVM Information Class
}
\details{
Instances of this class store SVM related information.
}
\section{Slots}{

\describe{
\item{\code{availPackages}}{installed SVM packages}

\item{\code{reqSVM}}{user requested SVM implementation}

\item{\code{reqPackage}}{user requested package}

\item{\code{reqSVMPar}}{user requested SVM parameters}

\item{\code{reqKernel}}{user requested kernel}

\item{\code{reqExplicit}}{user requested indictor of expl. rep. processing}

\item{\code{reqExplicitType}}{user requested expl. rep. type}

\item{\code{reqFeatureType}}{user requested feature type}

\item{\code{selSVM}}{selected SVM implementation}

\item{\code{selPackage}}{selected package}

\item{\code{selSVMPar}}{selected SVM parameters}

\item{\code{selKernel}}{selected kernel}

\item{\code{selExplicit}}{selected indictor of expl. rep. processing}

\item{\code{explicitKernel}}{kernel for explicit representation}

\item{\code{featureWeights}}{indicator for feature weights}

\item{\code{weightLimit}}{cutoff value for feature weights}

\item{\code{probModel}}{indicator for probability model}
}}

