% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{SymmetricPairKernel-class}
\alias{SymmetricPairKernel}
\alias{SymmetricPairKernel-class}
\alias{class:SymmetricPairKernel}
\title{Symmetric Pair Kernel Class}
\description{
Symmetric Pair Kernel Class
}
\details{
Instances of this class represent a kernel object for the
symmetric pair kernel. The kernel does not compute
similarity between single samples but between two pairs of
samples based on a regular sequence kernel for single
samples. The class is derived from \link{SequenceKernel}.
}
\section{Slots}{

\describe{
\item{\code{siKernel}}{single instance kernel}

\item{\code{kernelType}}{type of pair kernel}

\item{\code{r}}{exponent (for details see \link{gappyPairKernel})}
}}

