% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{KernelMatrix-class}
\alias{KernelMatrix}
\alias{KernelMatrix-class}
\alias{class:KernelMatrix}
\title{Kernel Matrix Class}
\description{
Kernel Matrix Class
}
\details{
Instances of this class are used in KeBABS for storing a
kernel matrix. The hidden data part ".Data" contains the
matrix.
}
\author{
Johannes Palme <kebabs@bioinf.jku.at>
}
\references{
\url{http://www.bioinf.jku.at/software/kebabs}\cr\cr J.
Palme, S. Hochreiter, and U. Bodenhofer (2015) KeBABS: an R
package for kernel-based analysis of biological sequences.
\emph{Bioinformatics} (accepted). DOI:
\href{http://dx.doi.org/10.1093/bioinformatics/btv176}{10.1093/bioinformatics/btv176}.
}

