% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{KernelMatrix-class}
\alias{KernelMatrix}
\alias{KernelMatrix-class}
\alias{class:KernelMatrix}
\title{Kernel Matrix Class}
\description{
Kernel Matrix Class
}
\details{
Instances of this class are used in KeBABS for storing a kernel matrix.
The hidden data part ".Data" contains the matrix.
}
\author{
Johannes Palme
}
\references{
\url{https://github.com/UBod/kebabs}\cr\cr
J. Palme, S. Hochreiter, and U. Bodenhofer (2015) KeBABS: an R package
for kernel-based analysis of biological sequences.
\emph{Bioinformatics}, 31(15):2574-2576.
DOI: \doi{10.1093/bioinformatics/btv176}.
}
