% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PredictionProfile-class}
\alias{PredictionProfile}
\alias{PredictionProfile-class}
\alias{class:PredictionProfile}
\title{Prediction Profile Class}
\description{
Prediction Profile Class
}
\details{
This class stores prediction profiles generated for a set of biological
sequences from a trained model. Prediction profiles show the relevance
of individual sequence positions for the prediction result.
}
\section{Slots}{

\describe{
\item{\code{sequences}}{sequence information for the samples with profiles}

\item{\code{baselines}}{baselines generated from the offset in the model spread\cr
to all sequence positions}

\item{\code{profiles}}{prediction profile information stored as dense matrix with\cr
the rows as samples and the columns as positions in the sample}

\item{\code{kernel}}{kernel used for training the model on which these prediction\cr
      profiles are based}
}}
\author{
Johannes Palme <kebabs@bioinf.jku.at>
}
\references{
\url{http://www.bioinf.jku.at/software/kebabs}\cr\cr
J. Palme, S. Hochreiter, and U. Bodenhofer (2015) KeBABS: an R package
for kernel-based analysis of biological sequences.
\emph{Bioinformatics}, 31(15):2574-2576, 2015.
DOI: \href{http://dx.doi.org/10.1093/bioinformatics/btv176}{10.1093/bioinformatics/btv176}.
}

