% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ROCData-class}
\alias{ROCData}
\alias{ROCData-class}
\alias{class:ROCData}
\title{ROC Data Class}
\description{
ROC Data Class
}
\details{
This class stores receiver operating characteristics (ROC) data.
}
\section{Slots}{

\describe{
\item{\code{AUC}}{area under ROC curve}

\item{\code{TPR}}{true positive rate for varying threshold}

\item{\code{FPR}}{false positive rate for varying threshold}
}}
\author{
Johannes Palme
}
\references{
\url{https://github.com/UBod/kebabs}\cr\cr
J. Palme, S. Hochreiter, and U. Bodenhofer (2015) KeBABS: an R package
for kernel-based analysis of biological sequences.
\emph{Bioinformatics}, 31(15):2574-2576.
DOI: \doi{10.1093/bioinformatics/btv176}.
}
