% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{SequenceKernel-class}
\alias{SequenceKernel}
\alias{SequenceKernel-class}
\alias{class:SequenceKernel}
\title{Sequence Kernel Class}
\description{
Sequence Kernel Class
}
\details{
This class represents the parent class for all sequence
kernels. It is an abstract class and must not be
instantiated.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{the kernel function is stored in this slot. It is executed
when the kernel matrix is created through invoking the kernel object.}
}}

