% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{SpectrumKernel-class}
\alias{SpectrumKernel}
\alias{SpectrumKernel-class}
\alias{class:SpectrumKernel}
\title{Spectrum Kernel Class}
\description{
Spectrum Kernel Class
}
\details{
Instances of this class represent a kernel object for the
spectrum kernel. The class is derived from
\link{SequenceKernel}.
}
\section{Slots}{

\describe{
\item{\code{k}}{length of the substrings considered by the kernel}

\item{\code{r}}{exponent (for details see \link{spectrumKernel})}

\item{\code{annSpec}}{when set the kernel evaluates annotation information}

\item{\code{distWeight}}{distance weighting function or vector}

\item{\code{normalized}}{data generated with this kernel object is normalized}

\item{\code{exact}}{use exact character set for evaluation}

\item{\code{ignoreLower}}{ignore lower case characters in the sequence}

\item{\code{presence}}{consider only the presence of kmers not their counts}

\item{\code{revComplement}}{consider a kmer and its reverse complement
as the same feature}

\item{\code{mixCoef}}{mixing coefficients for mixture kernel}
}}

